fix: week 3

+++ src/main/java/com/vasquez/mscredit/service/CreditCardService.java
@@ -1,22 +1,27 @@
 package com.vasquez.mscredit.service;
 
-import com.vasquez.mscredit.entity.Credit;
 import com.vasquez.mscredit.entity.CreditCard;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+/**
+ * Credit card service interface.
+ *
+ * @author Vasquez
+ * @version 1.0.
+ */
 public interface CreditCardService {
 
-    Mono<CreditCard> create(CreditCard request);
+  Mono<CreditCard> create(CreditCard request);
 
-    Mono<CreditCard> update(CreditCard request, String creditCardId);
+  Mono<CreditCard> update(CreditCard request, String creditCardId);
 
-    Mono<CreditCard> findById(String creditCardId);
+  Mono<CreditCard> findById(String creditCardId);
 
-    Flux<CreditCard> findAll();
+  Flux<CreditCard> findAll();
 
-    Mono<Void> deleteById(String creditCardId);
+  Mono<Void> deleteById(String creditCardId);
 
-    Mono<Integer> countByClientId(String clientId);
+  Mono<Long> countByClientId(String clientId);
 
 }
